requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_ipyw_jlab_nb_ext_conf=0.1.0=py37_0
anaconda-client=1.9.0=py37haa95532_0
anaconda-navigator=1.9.7=py37_0
anyio=3.5.0=py37haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h2bbff1b_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pypi_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
backports=1.1=pyhd3eb1b0_0
backports.functools_lru_cache=1.6.4=pyhd3eb1b0_0
backports.tempfile=1.0=pyhd3eb1b0_1
backports.weakref=1.0.post1=py_1
beautifulsoup4=4.10.0=pyh06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py37h080aedc_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2022.4.26=haa95532_0
certifi=2022.5.18.1=py37haa95532_0
cffi=1.15.0=py37h2bbff1b_1
chardet=4.0.0=py37haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37haa95532_0
clyent=1.2.2=py37_1
colorama=0.4.4=pyhd3eb1b0_0
conda=4.13.0=py37haa95532_0
conda-build=3.21.9=py37haa95532_0
conda-env=2.6.0=1
conda-package-handling=1.8.1=py37h8cc25b3_0
conda-verify=3.4.2=py_1
console_shortcut=0.1.1=3
cryptography=36.0.0=py37h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py37hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
filelock=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py37_1
glob2=0.7=pyhd3eb1b0_0
grad-cam=1.3.9=pypi_0
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py37haa95532_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.9.1=py37haa95532_0
ipython=7.31.1=py37haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37haa95532_1
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.2=py37haa95532_0
jupyter_server=1.13.5=pyhd3eb1b0_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.10.3=pyhd3eb1b0_1
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
jupyterthemes=0.20.0=py_1
kiwisolver=1.3.2=py37hd77b12b_0
lesscpy=0.15.0=pyhd8ed1ab_0
libarchive=3.4.2=h5e25573_0
libiconv=1.15=h1df5818_7
liblief=0.11.5=hd77b12b_1
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
libxml2=2.9.12=h0ad7f3c_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.0.1=py37h2bbff1b_0
matplotlib=3.4.3=py37haa95532_0
matplotlib-base=3.4.3=py37h49ac443_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
menuinst=1.4.16=py37he774522_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
munkres=1.1.4=py_0
navigator-updater=0.2.1=py37_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.4.4=py37haa95532_0
nbformat=5.3.0=py37haa95532_0
nest-asyncio=1.5.5=py37haa95532_0
notebook=6.4.8=py37haa95532_0
numexpr=2.8.1=py37hb80d3ca_0
numpy=1.21.5=py37h7a0a035_1
numpy-base=1.21.5=py37hca35cd5_1
opencv=3.4.2=py37h40b0b35_0
opencv-python=4.5.5.64=pypi_0
openssl=1.1.1o=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h6214cd6_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py37hdc2b20a_0
pip=21.2.4=py37haa95532_0
pkginfo=1.8.2=pyhd3eb1b0_0
ply=3.11=py_1
powershell_shortcut=0.0.1=2
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.8.0=py37h2bbff1b_1
py-lief=0.11.5=py37hd77b12b_1
py-opencv=3.4.2=py37hc319ecb_0
pycodestyle=2.8.0=pypi_0
pycosat=0.6.3=py37h2bbff1b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37hd77b12b_6
pyrsistent=0.18.0=py37h196d8e1_0
pysocks=1.7.1=py37_1
python=3.7.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python-libarchive-c=2.9=pyhd3eb1b0_1
python_abi=3.7=2_cp37m
pytorch=1.11.0=py3.7_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywin32=302=py37h827c3e9_1
pywinpty=2.0.2=py37h5da7b33_0
pyyaml=6.0=py37h2bbff1b_1
pyzmq=22.3.0=py37hd77b12b_2
qt=5.9.7=vc14h73c81de_0
qtpy=2.0.1=pyhd3eb1b0_0
requests=2.27.1=pyhd3eb1b0_0
ruamel_yaml=0.15.100=py37h2bbff1b_0
scikit-learn=1.0.2=py37hf11a4ad_1
scipy=1.7.3=py37h0a974cb_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py37haa95532_0
sip=4.19.13=py37hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py37haa95532_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.2=h2bbff1b_0
terminado=0.13.1=py37haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
toml=0.10.2=pypi_0
torchaudio=0.11.0=py37_cu113
torchsummary=1.5.1=pypi_0
torchvision=0.12.0=py37_cu113
tornado=6.1=py37h2bbff1b_0
tqdm=4.63.0=pyhd3eb1b0_0
traitlets=5.1.1=pyhd3eb1b0_0
ttach=0.0.3=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37haa95532_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37haa95532_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
winpty=0.4.3=4
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=hbd8134f_5
zstd=1.4.9=h19a0ad4_0
